+++ rust/poker_eval/src/web/analyzer.rs
@@ -1,5 +1,5 @@
 use std::{cmp::{self}, mem};
-use crate::{PokerError, get_unused_card, add_eval_card, set_used_card, HoleCards};
+use crate::{PokerError, get_unused_card, add_eval_card, set_used_card, HoleCards, PreflopPlayerInfo, PlayerPreFlopState, PlayerFlopResults, FlopSimulationResults, eval_current_draws, get_all_player_hole_cards, eval_current};
 use itertools::Itertools;
 use log::{debug, error, info, trace, warn};
 use postflop_solver::card_pair_to_index;
@@ -15,114 +15,6 @@ use wasm_bindgen::{prelude::wasm_bindgen, JsValue};
 type ResultType = u32;
 use serde::Serialize;
 
-//#[wasm_bindgen]
-#[derive(Default)]
-pub struct RankResults {
-    num_iterations: ResultType,
-
-    //win = 1, tie = 1 / num players in tie, loss = 0
-    win_eq: f64,
-    tie_eq: f64,
-
-    rank_family_count: [ResultType; 9],
-}
-
-#[derive(Default)]
-#[wasm_bindgen]
-pub struct PlayerFlopResults {
-    /*
-    This is when evaluating the flop vs the players
-    */
-    //pub num_iterations: ResultType,
-    player_index: usize,
-
-    //hand rankings as of the flop
-    //3 of them, flop turn river
-    street_rank_results: [RankResults; 3],
-
-    //turn & river
-    street_draws: [Draws; 2],
-}
-
-#[wasm_bindgen]
-impl PlayerFlopResults {
-    pub fn new() -> Self {
-        let d = Self::default();
-        //d.num_iterations = 0;
-        d
-    }
-
-    
-}
-
-
-#[wasm_bindgen]
-#[derive(Default, Serialize)]
-pub struct Draws {
-    pub num_iterations: ResultType,
-
-    pub gut_shot: ResultType,
-    pub str8_draw: ResultType,
-    pub flush_draw: ResultType,
-    pub backdoor_flush_draw: ResultType,
-    
-    pub one_overcard: ResultType,
-    pub two_overcards: ResultType,
-    
-    pub lo_paired: ResultType,
-    pub hi_paired: ResultType,
-    pub pp_paired: ResultType,
-}
-
-#[wasm_bindgen]
-impl Draws {
-    pub fn new() -> Self {
-        Self {
-            gut_shot: 0,
-            str8_draw: 0,
-            flush_draw: 0,
-            backdoor_flush_draw: 0,
-            one_overcard: 0,
-            two_overcards: 0,
-
-            lo_paired: 0,
-            hi_paired: 0,
-            pp_paired: 0,
-
-            num_iterations: 0,
-        }
-    }
-}
-
-#[derive(Eq, PartialEq, Debug)]
-#[repr(u8)]
-enum PlayerPreFlopState {
-    Disabled = 0,
-    UseHoleCards = 1,
-    UseRange = 2,
-}
-
-impl From<u8> for PlayerPreFlopState {
-    fn from(value: u8) -> Self {
-        unsafe { mem::transmute(cmp::min(value, 2)) }
-    }
-}
-
-impl Default for PlayerPreFlopState {
-    fn default() -> Self {
-        Self::Disabled
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Default)]
-pub struct PreflopPlayerInfo {
-    range_string: String,
-    //results: Results,
-    hole_cards: Option<HoleCards>,
-    range_set: InRangeType,
-    state: PlayerPreFlopState,
-}
 
 #[wasm_bindgen]
 //doing this to stop warnings in vs code about camel case in the wasm function names
@@ -133,83 +25,10 @@ pub struct flop_analyzer {
 }
 
 
-#[wasm_bindgen]
-pub struct FlopSimulationResults {
-    //Because of https://stackoverflow.com/questions/68243940/rust-wasm-bindgen-struct-with-string
-    //we don't want copy but we have accessors
-    all_villians: PlayerFlopResults,
-    flop_results: Vec<PlayerFlopResults>,
-    
-}
-
-#[wasm_bindgen]
-impl FlopSimulationResults {
-
-    //note these player_indexes are the index of active players
-
-    // The rust struct can't get passed via the worker interface, so we need primitive accessors
-    pub fn get_perc_family(&self, active_player_index: Option<usize>, street_index: usize, family_index: usize) -> f64 {
-        let r = if let Some(p_idx) = active_player_index {
-             &self.flop_results[p_idx].street_rank_results[street_index]
-        } else {
-            &self.all_villians.street_rank_results[street_index]
-        };
-          
-        r.rank_family_count[family_index] as f64 / r.num_iterations as f64
-    }
-    pub fn get_perc_family_or_better(&self, active_player_index: Option<usize>, street_index: usize, family_index: usize) -> f64 {
-        let mut total = 0.0;
-        for i in family_index..NUM_RANK_FAMILIES {
-            total += self.get_perc_family(active_player_index, street_index, i)
-        }
-        total
-    }
-    pub fn get_equity(&self, active_player_index: Option<usize>, street_index: usize) -> f64 {
-        
-        let r = if let Some(p_idx) = active_player_index {
-             &self.flop_results[p_idx].street_rank_results[street_index]
-        } else {
-            &self.all_villians.street_rank_results[street_index]
-        };
-
-        (r.win_eq + r.tie_eq) / r.num_iterations as f64
-    }
-
-    //This guy has no arrays, so we can just convert it to json
-    pub fn get_street_draw(&self, active_player_index: Option<usize>, draw_index: usize) -> Result<JsValue, JsValue> {
-        info!(
-            "get_street_draw: {} ",
-            draw_index
-        );
-        
-        Ok(serde_wasm_bindgen::to_value(
-        if let Some(p_idx) = active_player_index {            
-                &self.flop_results[p_idx].street_draws[draw_index]            
-        } else {
-            &self.all_villians.street_draws[draw_index]
-        })?)
-        
-    }
-
-    pub fn get_num_players(&self) -> usize {
-        self.flop_results.len()
-    }
-
-    //Convert from active_player_index to original
-    pub fn get_player_index(&self, player_index: usize) -> usize {
-        info!(
-            "get_player_index: {} ",
-            player_index
-        );
-        
-        self.flop_results[player_index].player_index
-    }
-}
-
 //hero is 0
-const MAX_PLAYERS: usize = 5;
+pub const MAX_PLAYERS: usize = 5;
 
-const MAX_RAND_NUMBER_ATTEMPS: usize = 1000;
+pub const MAX_RAND_NUMBER_ATTEMPS: usize = 1000;
 
 #[wasm_bindgen]
 impl flop_analyzer {
@@ -579,289 +398,6 @@ impl flop_analyzer {
 
 }
 
-/*
-Assumes all players have either hole cards or their ranges chosen
-*/
-pub fn eval_current(
-    active_players: &[(usize, &PreflopPlayerInfo)],
-    player_cards: &[HoleCards],
-    eval_cards: &mut Vec<Card>,
-    flop_results: &mut Vec<PlayerFlopResults>,
-    //treat first active player as the hero, all others as villians
-    villian_results: &mut PlayerFlopResults,
-    street_index: usize,
-) -> Result<(), PokerError> {
-    if eval_cards.len() < 3 {
-        return Err(PokerError::from_string(format!(
-            "eval_current: eval_cards needs at least 3 cards, but had {} cards",
-            eval_cards.len()
-        )));
-    }
-    if eval_cards.len() > 5 {
-        return Err(PokerError::from_string(format!(
-            "eval_current: too many eval_cards, should be 5 max, but had {} cards",
-            eval_cards.len()
-        )));
-    }
-
-    let n_players = active_players.len();
-    assert!(n_players > 1);
-    assert_eq!(player_cards.len(), n_players);
-
-    let mut hand_evals: Vec<Rank> = Vec::with_capacity(n_players);
-
-    for (active_index, (_p_idx, p)) in active_players.iter().enumerate() {
-        assert!(p.state != PlayerPreFlopState::Disabled);
-
-        //For players with ranges we already chose their cards
-
-        player_cards[active_index].add_to_eval(eval_cards);
-        
-        let rank = rank_cards(&eval_cards);
-
-        update_results_from_rank(
-            &mut flop_results[active_index].street_rank_results[street_index],
-            rank,
-        );
-
-        hand_evals.push(rank);
-
-        player_cards[active_index].remove_from_eval(eval_cards)?;
-    }
-
-    //Best villian hand
-    let best_villian_rank = hand_evals[1..]
-        .iter()
-        .fold(Rank::HighCard(0), |acc, &x| acc.max(x));
-    update_results_from_rank(&mut villian_results.street_rank_results[street_index], best_villian_rank);
-
-    let winner_indexes = indices_of_max_values(&hand_evals);
-
-    assert!(winner_indexes.len() > 0);
-
-    for winner_idx in winner_indexes.iter() {
-        let results = &mut flop_results[*winner_idx].street_rank_results[street_index];
-        if winner_indexes.len() == 1 {
-            results.win_eq += 1.0;
-            if *winner_idx > 0 {
-                villian_results.street_rank_results[street_index].win_eq += 1.0;
-            }
-        } else {
-            results.tie_eq += 1.0 / winner_indexes.len() as f64;
-
-            if *winner_idx > 0 {
-                villian_results.street_rank_results[street_index].tie_eq += 1.0 / winner_indexes.len() as f64;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-pub fn eval_current_draws(
-    active_players: &[(usize, &PreflopPlayerInfo)],
-    player_cards: &[HoleCards],
-    eval_cards: &Vec<Card>,
-    flop_results: &mut Vec<PlayerFlopResults>,
-    //treat first active player as the hero, all others as villians
-    villian_results: &mut PlayerFlopResults,
-    draw_index: usize,
-) -> Result<(), PokerError> {
-    if eval_cards.len() < 3 {
-        return Err(PokerError::from_string(format!(
-            "eval_current: eval_cards needs at least 3 cards, but had {} cards",
-            eval_cards.len()
-        )));
-    }
-    if eval_cards.len() >= 5 {
-        return Err(PokerError::from_string(format!(
-            "eval_current: too many eval_cards, should be 4 max since we are drawing, but had {} cards",
-            eval_cards.len()
-        )));
-    }
-
-    //flop  = 0
-    //turn = 1
-    //we don't draw on the river
-    assert!(draw_index < 2);
-
-    let n_players = active_players.len();
-    assert!(n_players > 1);
-    assert_eq!(player_cards.len(), n_players);
-
-    let mut best_villian_draws = PartialRankContainer::default();
-
-    for (active_index, (_p_idx, p)) in active_players.iter().enumerate() {
-        assert!(p.state != PlayerPreFlopState::Disabled);
-
-        //For players with ranges we already chose their cards
-
-        let hc = &player_cards[active_index];
-        let prc = partial_rank_cards(
-            &hc, 
-            &eval_cards);
-
-        update_draw(
-            &mut flop_results[active_index].street_draws[draw_index],
-            &prc,
-        );
-
-        if active_index > 0 {
-            best_villian_draws.merge_best(&prc);
-        }
-
-    }
-
-    update_draw(
-        &mut villian_results.street_draws[draw_index],
-        &best_villian_draws,
-    );
-
-    Ok(())
-}
-
-
-fn get_all_player_hole_cards(
-    active_players: &[(usize, &PreflopPlayerInfo)],
-    rng: &mut StdRng,
-    cards_used: &mut CardUsedType,
-) -> Result<Vec<HoleCards>, PokerError> {
-    let mut player_cards: Vec<HoleCards> = Vec::with_capacity(active_players.len());
-
-    
-    for (p_idx, p) in active_players.iter() {
-        assert!(p.state != PlayerPreFlopState::Disabled);
-
-        if p.state == PlayerPreFlopState::UseHoleCards {
-            let pc = p.hole_cards.ok_or(PokerError::from_string(format!("Player missing hole cards")))?;
-            player_cards.push(pc);
-            continue;
-        }
-
-        assert_eq!(p.state, PlayerPreFlopState::UseRange);
-
-        //Now deal with ranges
-        let mut attempts = 0;
-        let mut card1_index;
-        let mut card2_index;
-
-        loop {
-            attempts += 1;
-
-            if attempts > MAX_RAND_NUMBER_ATTEMPS {
-                return Err(PokerError::from_string(
-                    format!("Unable to find cards for player {} after {} attempts.  Cards used count {} range str {} == {:.1}%",
-                    p_idx, attempts, cards_used.count_ones(),
-                    &p.range_string, p.range_set.count_ones() as f64 / 2652.0 * 100.0)
-                ));
-            }
-
-            card1_index = get_unused_card(rng, cards_used).unwrap();
-            card2_index = get_unused_card(rng, cards_used).unwrap();
-
-            if card1_index == card2_index {
-                continue;
-            }
-
-            let range_index = card_pair_to_index(card1_index as u8, card2_index as u8);
-
-            if !p.range_set[range_index] {
-                continue;
-            }
-
-            break;
-        }
-
-        //we set their cards
-        let pc = HoleCards::new(
-            Card::try_from(card1_index)?, Card::try_from(card2_index)?)?;
-        pc.set_used(cards_used)?;
-        player_cards.push(pc);
-    }
-
-    Ok(player_cards)
-}
-
-
-fn update_results_from_rank(results: &mut RankResults, rank: Rank) {
-    results.num_iterations += 1;
-    results.rank_family_count[rank.get_family_index()] += 1;
-}
-
-fn update_draw(results: &mut Draws, prc: &PartialRankContainer) {
-
-    results.num_iterations += 1;
-
-    if let Some(sd) = prc.straight_draw.as_ref() {
-        match sd.straight_draw_type {
-            StraightDrawType::GutShot(_cv) => results.gut_shot += 1,
-            StraightDrawType::OpenEnded => results.str8_draw += 1,
-            StraightDrawType::DoubleGutShot => results.str8_draw += 1,
-        }
-    }
-
-    if let Some(fd) = prc.flush_draw.as_ref() {
-        match fd.flush_draw_type
-        {
-            FlushDrawType::FlushDraw => results.flush_draw += 1,
-            FlushDrawType::BackdoorFlushDraw => results.backdoor_flush_draw += 1,
-        }
-    }
-
-    if let Some(_pp) = prc.pocket_pair.as_ref() {
-        results.pp_paired += 1;
-    }
-
-    if let Some(_hi) = prc.hi_pair.as_ref() {
-        results.hi_paired += 1;
-    }
-
-    if let Some(_lo) = prc.lo_pair.as_ref() {
-        results.lo_paired += 1;
-    }
-
-    let mut num_overcards: usize = 0;
-
-    if let Some (hi) = prc.hi_card.as_ref() {
-        if hi.number_above == 0 {
-            num_overcards += 1;
-
-            //Only count lower card as overcard if the higher one did too
-            if let Some (lo) = prc.lo_card.as_ref() {
-                if lo.number_above == 0 {
-                    num_overcards += 1;
-                }
-            }
-        }
-    }
-
-    assert!(num_overcards <= 2);
-
-    if num_overcards == 1 {
-        results.one_overcard += 1;
-    } else if num_overcards == 2 {
-        results.two_overcards += 1;
-    }
-}
-
-//returns winners and how many players were considered (non None rank)
-fn indices_of_max_values(arr: &[Rank]) -> Vec<usize> {
-    
-    let mut max_indices = Vec::with_capacity(MAX_PLAYERS);
-    let mut max_value = Rank::HighCard(0);
-
-    for (index, &value) in arr.iter().enumerate() {
-        if value > max_value {
-            max_value = value;
-            max_indices.clear();
-            max_indices.push(index);
-        } else if value == max_value {
-            max_indices.push(index);
-        }
-    }
-
-    max_indices
-}
 
 
 #[cfg(test)]
