+++ rust/poker_eval/src/web/mod.rs
@@ -1,2 +1,17 @@
 mod analyzer;
 pub use analyzer::*;
+
+mod draws;
+pub use draws::*;
+
+mod simul_results;
+pub use simul_results::*;
+
+mod player_results;
+pub use player_results::*;
+
+mod ranks;
+pub use ranks::*;
+
+mod preflop_player_info;
+pub use preflop_player_info::*;
