+++ vue-poker/public/build_info.txt
@@ -1,2 +1,1 @@
-Built: 2023/12/18 17:13:41
-Commit Hash: Local
+{"build_date":1703010844482,"github_sha":"Local"}
